refactor: remove TeaGrid and Cart components, update import paths in TeaList, and enhance TeaProductCard with improved structure and comments for better readability
