--- src/pages/Cart.jsx
@@ -1,54 +1,0 @@
-import React from 'react';
-import { useCart } from '../context/CartContext';
-
-const Cart = () => {
-  const { cart, removeFromCart, updateQuantity } = useCart();
-
-  const total = cart.reduce((sum, item) => sum + item.price * item.quantity, 0);
-
-  return (
-    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
-      <h1 className="text-3xl font-bold text-gray-900 mb-8">Your Cart</h1>
-      {cart.length === 0 ? (
-        <p className="text-gray-500">Your cart is empty</p>
-      ) : (
-        <div className="space-y-8">
-          {cart.map((item) => (
-            <div key={item.id} className="flex items-center justify-between border-b pb-4">
-              <div className="flex items-center space-x-4">
-                <img src={item.image} alt={item.name} className="w-16 h-16 object-cover rounded" />
-                <div>
-                  <h3 className="text-lg font-medium text-gray-900">{item.name}</h3>
-                  <p className="text-gray-500">${item.price}</p>
-                </div>
-              </div>
-              <div className="flex items-center space-x-4">
-                <input
-                  type="number"
-                  min="1"
-                  value={item.quantity}
-                  onChange={(e) => updateQuantity(item.id, parseInt(e.target.value))}
-                  className="w-16 px-2 py-1 border rounded"
-                />
-                <button
-                  onClick={() => removeFromCart(item.id)}
-                  className="text-red-600 hover:text-red-800"
-                >
-                  Remove
-                </button>
-              </div>
-            </div>
-          ))}
-          <div className="flex justify-between items-center pt-4">
-            <span className="text-xl font-medium">Total: ${total.toFixed(2)}</span>
-            <button className="bg-green-600 text-white px-6 py-2 rounded hover:bg-green-700">
-              Checkout
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default Cart;
